refactor: ajustar estrutura do componente Blog com animações e modularização
